+++ hi-ml-azure/src/health/azure/__init__.py
@@ -5,7 +5,7 @@
 
 from health.azure.azure_util import (fetch_run, set_environment_variables_for_multi_node, split_recovery_id,
                                      get_most_recent_run, download_files_from_run_id, download_checkpoints_from_run_id,
-                                     download_from_datastore, upload_to_datastore)
+                                     download_from_datastore, upload_to_datastore, torch_barrier)
 from health.azure.datasets import DatasetConfig
 from health.azure.himl import (AzureRunInfo, create_run_configuration, create_script_run, get_workspace, submit_run,
                                submit_to_azure_if_needed)
@@ -24,4 +24,5 @@ __all__ = ["fetch_run",
            "download_files_from_run_id",
            "download_checkpoints_from_run_id",
            "download_from_datastore",
-           "upload_to_datastore"]
+           "upload_to_datastore",
+           "torch_barrier"]