+++ CHANGELOG.md
@@ -22,10 +22,11 @@ created.
 
 ### Deprecated
 
-## 0.1.8 (2021-10-05)
+## 0.1.8 (2021-10-06)
 
 ### Added
 - ([#123](https://github.com/microsoft/hi-ml/pull/123)) Add helper function to download checkpoint files
+- ([#128](https://github.com/microsoft/hi-ml/pull/128)) When downloading files in a distributed PyTorch job, a barrier is used to synchronize the processes.
 
 ### Changed
 - ([#127](https://github.com/microsoft/hi-ml/pull/127)) The field `is_running_in_azure` of `AzureRunInfo` has been renamed to `is_running_in_azure_ml`